+++ components/text-editor/plugins/MentionPlugin/MentionPlugin.tsx
@@ -1,4 +1,10 @@
-import React, { useState, useMemo, useCallback } from "react";
+import React, {
+	useState,
+	useMemo,
+	useCallback,
+	useRef,
+	useEffect,
+} from "react";
 import * as ReactDOM from "react-dom";
 import { TextNode } from "lexical";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
@@ -10,6 +16,7 @@ import {
 } from "@lexical/react/LexicalTypeaheadMenuPlugin";
 
 import { $createMentionNode } from "./MentionNode";
+import { UserIcon } from "../../icons";
 
 const PUNCTUATION =
 	"\\.,\\+\\*\\?\\$\\@\\|#{}\\(\\)\\^\\-\\[\\]\\\\/!%'\"~=<>_:;";
@@ -59,8 +66,6 @@ const AtSignMentionsRegexAliasRegex = new RegExp(
 		")$",
 );
 
-const SUGGESTION_LIST_LENGTH_LIMIT = 5;
-
 function checkForAtSignMentions(
 	text: string,
 	minMatchLength: number,
@@ -88,7 +93,7 @@ function checkForAtSignMentions(
 }
 
 function getPossibleQueryMatch(text: string): MenuTextMatch | null {
-	return checkForAtSignMentions(text, 1);
+	return checkForAtSignMentions(text, 0);
 }
 
 class MentionTypeaheadOption extends MenuOption {
@@ -140,34 +145,88 @@ function MentionsTypeaheadMenuItem({
 		</li>
 	);
 }
+
+function defaultLookupService(
+	mentionString: string | null,
+	mentionsData: Array<{
+		mentionName: string;
+		label: string;
+	}>,
+) {
+	const [results, setResults] = useState<
+		Array<{ mentionName: string; label: string }>
+	>([]);
+
+	function defaultSearch(
+		string: string,
+		callback: (results: Array<{ mentionName: string; label: string }>) => void,
+	): void {
+		setTimeout(() => {
+			const results = mentionsData.filter((mention) =>
+				mention.label.toLowerCase().includes(string.toLowerCase()),
+			);
+			callback(results);
+		}, 500);
+	}
+
+	useEffect(() => {
+		if (mentionString === null) {
+			setResults([]);
+		} else {
+			defaultSearch(mentionString, (newResults) => {
+				setResults(newResults);
+			});
+		}
+	}, [mentionString]);
+
+	return results;
+}
+
 type Props = {
-	useMentionLookupService: (mentionString: string | null) => Array<{
+	mentionsData: Array<{
+		mentionName: string;
+		label: string;
+	}>;
+	useMentionLookupService?: (
+		mentionString: string | null,
+		mentionsData: Array<{
+			mentionName: string;
+			label: string;
+		}>,
+	) => Array<{
 		mentionName: string;
 		label: string;
 	}>;
 };
+
 export default function NewMentionsPlugin({
+	mentionsData,
 	useMentionLookupService,
 }: Props): JSX.Element | null {
 	const [editor] = useLexicalComposerContext();
+	const dropDownRef = useRef<null | HTMLDivElement>(null);
 	const [searchquery, setSearchquery] = useState<string | null>(null);
-	const results = useMentionLookupService(searchquery);
+	const results = useMentionLookupService
+		? useMentionLookupService(searchquery, mentionsData)
+		: defaultLookupService(searchquery, mentionsData);
 	const checkForSlashTriggerMatch = useBasicTypeaheadTriggerMatch("/", {
 		minLength: 0,
 	});
 
 	const options = useMemo(
 		() =>
-			results
-				.map(
-					(result) =>
-						new MentionTypeaheadOption(
-							result.mentionName,
-							<i className="icon user" />,
-							result.label,
+			results.map(
+				(result) =>
+					new MentionTypeaheadOption(
+						result.mentionName,
+						(
+							<i className="icon">
+								<UserIcon />
+							</i>
 						),
-				)
-				.slice(0, SUGGESTION_LIST_LENGTH_LIMIT),
+						result.label,
+					),
+			),
 		[results],
 	);
 
@@ -218,32 +277,35 @@ export default function NewMentionsPlugin({
 				menuRenderFn={(
 					anchorElementRef,
 					{ selectedIndex, selectOptionAndCleanUp, setHighlightedIndex },
-				) =>
-					anchorElementRef.current && results.length
-						? ReactDOM.createPortal(
-								<div className="typeahead-popover mentions-menu">
-									<ul>
-										{options.map((option, i: number) => (
-											<MentionsTypeaheadMenuItem
-												index={i}
-												isSelected={selectedIndex === i}
-												onClick={() => {
-													setHighlightedIndex(i);
-													selectOptionAndCleanUp(option);
-												}}
-												onMouseEnter={() => {
-													setHighlightedIndex(i);
-												}}
-												key={option.key}
-												option={option}
-											/>
-										))}
-									</ul>
-								</div>,
-								anchorElementRef.current,
-						  )
-						: null
-				}
+				) => {
+					if (anchorElementRef.current && results.length) {
+						return ReactDOM.createPortal(
+							<div
+								className="typeahead-popover mentions-menu"
+								ref={dropDownRef}
+							>
+								<ul>
+									{options.map((option, i: number) => (
+										<MentionsTypeaheadMenuItem
+											index={i}
+											isSelected={selectedIndex === i}
+											onClick={() => {
+												setHighlightedIndex(i);
+												selectOptionAndCleanUp(option);
+											}}
+											onMouseEnter={() => {
+												setHighlightedIndex(i);
+											}}
+											key={option.key}
+											option={option}
+										/>
+									))}
+								</ul>
+							</div>,
+							anchorElementRef.current,
+						);
+					} else return null;
+				}}
 			/>
 		);
 	else {