+++ components/text-editor/src/Toolbar.tsx
@@ -2,51 +2,42 @@ import React from "react";
 import {
 	$isListNode,
 	INSERT_UNORDERED_LIST_COMMAND,
+	INSERT_ORDERED_LIST_COMMAND,
 	ListNode,
 	REMOVE_LIST_COMMAND,
 } from "@lexical/list";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import {
-	$createHeadingNode,
-	$isHeadingNode,
-	HeadingTagType,
-} from "@lexical/rich-text";
 import {
 	$findMatchingParent,
 	$getNearestNodeOfType,
 	mergeRegister,
 } from "@lexical/utils";
 import {
-	$createParagraphNode,
 	$getSelection,
 	$isRangeSelection,
 	$isRootOrShadowRoot,
 	CAN_REDO_COMMAND,
 	CAN_UNDO_COMMAND,
 	COMMAND_PRIORITY_CRITICAL,
-	DEPRECATED_$isGridSelection,
 	FORMAT_TEXT_COMMAND,
 	LexicalEditor,
 	REDO_COMMAND,
 	SELECTION_CHANGE_COMMAND,
 	UNDO_COMMAND,
 } from "lexical";
-import { $setBlocksType } from "@lexical/selection";
 import { useCallback, useEffect, useState } from "react";
 
 import useModal from "../utils/useModal";
-import { getSelectedNode } from "../utils/getSelectedNode";
-import { $isTableNode } from "../plugins/TablePlugin/TableNode";
 import { InsertNewTableDialog } from "../plugins/TablePlugin/TablePlugin";
 import { InsertImageDialog } from "../plugins/ImagePlugin/ImagePlugin";
 import {
 	BoldIcon,
 	BulletListIcon,
-	HeadingIcon,
 	ImageIcon,
 	ItalicIcon,
-	ParagraphIcon,
+	NumberedListIcon,
 	RedoIcon,
+	SaveIcon,
 	TableIcon,
 	UnderlineIcon,
 	UndoIcon,
@@ -54,63 +45,37 @@ import {
 
 const blockTypeToBlockName = {
 	bullet: "Bulleted List",
-	h1: "Heading 1",
 	paragraph: "Normal",
+	number: "Numbered List",
 };
 
-const rootTypeToRootName = {
-	root: "Root",
-	table: "Table",
-};
 function buttonActiveClass(active: boolean) {
 	if (active) return "active dropdown-item-active";
 	else return "";
 }
-const toolbarTextClass = (isActive: boolean) => {
-	return isActive ? "text-gray-700" : "text-gray-500";
-};
 
 function BlockTextFormat({
 	editor,
 	blockType,
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	rootType,
 	disabled = false,
+	showToolbarText,
 }: {
 	blockType: keyof typeof blockTypeToBlockName;
-	rootType: keyof typeof rootTypeToRootName;
 	editor: LexicalEditor;
 	disabled?: boolean;
+	showToolbarText?: boolean;
 }): JSX.Element {
-	const formatParagraph = () => {
-		editor.update(() => {
-			const selection = $getSelection();
-			if (
-				$isRangeSelection(selection) ||
-				DEPRECATED_$isGridSelection(selection)
-			) {
-				$setBlocksType(selection, () => $createParagraphNode());
-			}
-		});
-	};
-
-	const formatHeading = (headingSize: HeadingTagType) => {
-		if (blockType !== headingSize) {
-			editor.update(() => {
-				const selection = $getSelection();
-				if (
-					$isRangeSelection(selection) ||
-					DEPRECATED_$isGridSelection(selection)
-				) {
-					$setBlocksType(selection, () => $createHeadingNode(headingSize));
-				}
-			});
-		}
-	};
-
 	const formatBulletList = () => {
 		if (blockType !== "bullet") {
 			editor.dispatchCommand(INSERT_UNORDERED_LIST_COMMAND, undefined);
+		} else {
+			editor.dispatchCommand(REMOVE_LIST_COMMAND, undefined);
+		}
+	};
+
+	const formatNumberedList = () => {
+		if (blockType !== "number") {
+			editor.dispatchCommand(INSERT_ORDERED_LIST_COMMAND, undefined);
 		} else {
 			editor.dispatchCommand(REMOVE_LIST_COMMAND, undefined);
 		}
@@ -120,57 +85,45 @@ function BlockTextFormat({
 		<>
 			<button
 				disabled={disabled}
-				className={
-					"toolbar-item " + buttonActiveClass(blockType === "paragraph")
-				}
-				onClick={formatParagraph}
-			>
-				<i className="icon">
-					<ParagraphIcon />
-				</i>
-				<span className={"text " + toolbarTextClass(blockType === "paragraph")}>
-					Normal
-				</span>
-			</button>
-			<button
-				disabled={disabled}
-				className={"toolbar-item " + buttonActiveClass(blockType === "h1")}
-				onClick={() => formatHeading("h1")}
-			>
-				<i className="icon ">
-					<HeadingIcon />
-				</i>
-				<span className="text">Heading 1</span>
-			</button>
-			<button
-				disabled={disabled}
 				className={"toolbar-item " + buttonActiveClass(blockType === "bullet")}
 				onClick={formatBulletList}
 			>
 				<i className="icon ">
 					<BulletListIcon />
 				</i>
-				<span className="text">Bullet List</span>
+				{showToolbarText && <span className="text">Bullet List</span>}
+			</button>
+			<button
+				disabled={disabled}
+				className={"toolbar-item " + buttonActiveClass(blockType === "number")}
+				onClick={formatNumberedList}
+			>
+				<i className="icon ">
+					<NumberedListIcon />
+				</i>
+				{showToolbarText && <span className="text">Numbered List</span>}
 			</button>
 		</>
 	);
 }
 
 function Divider(): JSX.Element {
-	return <div className="w-[1px] bg-gray-300 mx-1" />;
+	return <div className="divider" />;
 }
+
 export default function ToolbarPlugin({
-	convertFileToImageUrl,
+	convertFilesToImageUrl,
+	setEditable,
+	showToolbarText,
 }: {
-	convertFileToImageUrl: (files: FileList | null) => string | null;
+	convertFilesToImageUrl: (files: FileList | null) => Array<string> | null;
+	setEditable?: React.Dispatch<React.SetStateAction<boolean>>;
+	showToolbarText: boolean;
 }): JSX.Element {
 	const [editor] = useLexicalComposerContext();
 	const [activeEditor, setActiveEditor] = useState(editor);
 	const [blockType, setBlockType] =
 		useState<keyof typeof blockTypeToBlockName>("paragraph");
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	const [rootType, setRootType] =
-		useState<keyof typeof rootTypeToRootName>("root");
 
 	const [isBold, setIsBold] = useState(false);
 	const [isItalic, setIsItalic] = useState(false);
@@ -178,7 +131,6 @@ export default function ToolbarPlugin({
 	const [canUndo, setCanUndo] = useState(false);
 	const [canRedo, setCanRedo] = useState(false);
 	const [modal, showModal] = useModal();
-	const [isEditable, setIsEditable] = useState(() => editor.isEditable());
 
 	const $updateToolbar = useCallback(() => {
 		const selection = $getSelection();
@@ -203,14 +155,6 @@ export default function ToolbarPlugin({
 			setIsItalic(selection.hasFormat("italic"));
 			setIsUnderline(selection.hasFormat("underline"));
 
-			const node = getSelectedNode(selection);
-			const tableNode = $findMatchingParent(node, $isTableNode);
-			if ($isTableNode(tableNode)) {
-				setRootType("table");
-			} else {
-				setRootType("root");
-			}
-
 			if (elementDOM !== null) {
 				if ($isListNode(element)) {
 					const parentList = $getNearestNodeOfType<ListNode>(
@@ -222,9 +166,7 @@ export default function ToolbarPlugin({
 						: element.getListType();
 					setBlockType(type);
 				} else {
-					const type = $isHeadingNode(element)
-						? element.getTag()
-						: element.getType();
+					const type = element.getType();
 					if (type in blockTypeToBlockName) {
 						setBlockType(type as keyof typeof blockTypeToBlockName);
 					}
@@ -248,7 +190,7 @@ export default function ToolbarPlugin({
 	useEffect(() => {
 		return mergeRegister(
 			editor.registerEditableListener((editable) => {
-				setIsEditable(editable);
+				setEditable && setEditable(editable);
 			}),
 			activeEditor.registerUpdateListener(({ editorState }) => {
 				editorState.read(() => {
@@ -275,116 +217,134 @@ export default function ToolbarPlugin({
 	}, [$updateToolbar, activeEditor, editor]);
 
 	return (
-		<div className="toolbar">
-			<button
-				disabled={!canUndo || !isEditable}
-				onClick={() => {
-					activeEditor.dispatchCommand(UNDO_COMMAND, undefined);
-				}}
-				className="toolbar-item spaced"
-				aria-label="Undo"
-			>
-				<i className="format ">
-					<UndoIcon />
-				</i>
-			</button>
-			<button
-				disabled={!canRedo || !isEditable}
-				onClick={() => {
-					activeEditor.dispatchCommand(REDO_COMMAND, undefined);
-				}}
-				className="toolbar-item spaced"
-				aria-label="Redo"
-			>
-				<i className="format ">
-					<RedoIcon />
-				</i>
-			</button>
-			<Divider />
-			{blockType in blockTypeToBlockName && activeEditor === editor && (
-				<>
-					<BlockTextFormat
-						disabled={!isEditable}
-						blockType={blockType}
-						rootType={rootType}
-						editor={editor}
-					/>
-					<Divider />
-				</>
+		<div className="toolbar__wrapper">
+			<div className="toolbar">
+				<button
+					disabled={!canUndo || !editor.isEditable()}
+					onClick={() => {
+						activeEditor.dispatchCommand(UNDO_COMMAND, undefined);
+					}}
+					className="toolbar-item spaced"
+					aria-label="Undo"
+				>
+					<i className="format ">
+						<UndoIcon />
+					</i>
+				</button>
+				<button
+					disabled={!canRedo || !editor.isEditable()}
+					onClick={() => {
+						activeEditor.dispatchCommand(REDO_COMMAND, undefined);
+					}}
+					className="toolbar-item spaced"
+					aria-label="Redo"
+				>
+					<i className="format ">
+						<RedoIcon />
+					</i>
+				</button>
+				<Divider />
+				{blockType in blockTypeToBlockName && activeEditor === editor && (
+					<>
+						<BlockTextFormat
+							blockType={blockType}
+							editor={editor}
+							disabled={!editor.isEditable()}
+							showToolbarText={showToolbarText}
+						/>
+						<Divider />
+					</>
+				)}
+				<button
+					disabled={!editor.isEditable()}
+					onClick={() => {
+						activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "bold");
+					}}
+					className={"toolbar-item spaced " + (isBold ? "active" : "")}
+					aria-label={`Format text as bold`}
+				>
+					<i className="format ">
+						<BoldIcon />
+					</i>
+				</button>
+				<button
+					disabled={!editor.isEditable()}
+					onClick={() => {
+						activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "italic");
+					}}
+					className={"toolbar-item spaced " + (isItalic ? "active" : "")}
+					aria-label={`Format text as Italic`}
+				>
+					<i className="format ">
+						<ItalicIcon />
+					</i>
+				</button>
+				<button
+					disabled={!editor.isEditable()}
+					onClick={() => {
+						activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "underline");
+					}}
+					className={"toolbar-item spaced " + (isUnderline ? "active" : "")}
+					aria-label={`Format text with underline`}
+				>
+					<i className="format ">
+						<UnderlineIcon />
+					</i>
+				</button>
+				<Divider />
+				<button
+					disabled={!editor.isEditable() || activeEditor !== editor}
+					className="toolbar-item"
+					onClick={() => {
+						showModal("Insert Table", (onClose) => (
+							<InsertNewTableDialog
+								activeEditor={activeEditor}
+								onClose={onClose}
+							/>
+						));
+					}}
+				>
+					<i className="icon ">
+						<TableIcon />
+					</i>
+					{showToolbarText && <span className="text">Table</span>}
+				</button>
+				<Divider />
+				<button
+					disabled={!editor.isEditable() || activeEditor !== editor}
+					className="toolbar-item"
+					onClick={() => {
+						showModal("Insert Image", (onClose) => (
+							<InsertImageDialog
+								activeEditor={activeEditor}
+								onClose={onClose}
+								convertFilesToImageUrl={convertFilesToImageUrl}
+							/>
+						));
+					}}
+				>
+					<i className="icon ">
+						<ImageIcon />
+					</i>
+					{showToolbarText && <span className="text">Image</span>}
+				</button>
+			</div>
+			{setEditable && (
+				<div className="toolbar">
+					<button
+						disabled={!editor.isEditable()}
+						className="toolbar-item"
+						onClick={() => {
+							setEditable && setEditable(false);
+						}}
+					>
+						<i className="icon ">
+							<SaveIcon />
+						</i>
+						{showToolbarText && <span className="text">Save</span>}
+					</button>
+				</div>
 			)}
-			<button
-				disabled={!isEditable}
-				onClick={() => {
-					activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "bold");
-				}}
-				className={"toolbar-item spaced " + (isBold ? "active" : "")}
-				aria-label={`Format text as bold`}
-			>
-				<i className="format ">
-					<BoldIcon />
-				</i>
-			</button>
-			<button
-				disabled={!isEditable}
-				onClick={() => {
-					activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "italic");
-				}}
-				className={"toolbar-item spaced " + (isItalic ? "active" : "")}
-				aria-label={`Format text as Italic`}
-			>
-				<i className="format ">
-					<ItalicIcon />
-				</i>
-			</button>
-			<button
-				disabled={!isEditable}
-				onClick={() => {
-					activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, "underline");
-				}}
-				className={"toolbar-item spaced " + (isUnderline ? "active" : "")}
-				aria-label={`Format text with underline`}
-			>
-				<i className="format ">
-					<UnderlineIcon />
-				</i>
-			</button>
-			<Divider />
-			<button
-				disabled={!isEditable}
-				className="toolbar-item"
-				onClick={() => {
-					showModal("Insert Table", (onClose) => (
-						<InsertNewTableDialog
-							activeEditor={activeEditor}
-							onClose={onClose}
-						/>
-					));
-				}}
-			>
-				<i className="icon ">
-					<TableIcon />
-				</i>
-				<span className="text">Table (Experimental)</span>
-			</button>
-			<Divider />
-			<button
-				disabled={!isEditable}
-				className="toolbar-item "
-				onClick={() => {
-					showModal("Insert Image", (onClose) => (
-						<InsertImageDialog
-							activeEditor={activeEditor}
-							onClose={onClose}
-							convertFileToImageUrl={convertFileToImageUrl}
-						/>
-					));
-				}}
-			>
-				<i className="icon ">
-					<ImageIcon />
-				</i>
-				<span className="text">Image</span>
-			</button>
 			{modal}
 		</div>
 	);