+++ components/text-editor/src/index.tsx
@@ -14,28 +14,49 @@ import { TableContext } from "../plugins/TablePlugin/TablePlugin";
 import { ImageNode } from "../plugins/ImagePlugin/ImageNode";
 
 type Props = {
-	useMentionLookupService: (mentionString: string | null) => Array<{
+	mentionsData?:
+		| Array<{
+				mentionName: string;
+				label: string;
+		  }>
+		| false;
+	useMentionLookupService?: (
+		mentionString: string | null,
+		mentionsData: Array<{
+			mentionName: string;
+			label: string;
+		}>,
+	) => Array<{
 		mentionName: string;
 		label: string;
 	}>;
-	convertFileToImageUrl: (files: FileList | null) => string | null;
+	convertFilesToImageUrl: (files: FileList | null) => Array<string> | null;
 	onChangeCallback?: (editorRef: LexicalEditor | null, payload: any) => void;
 	wrapperClass?: string;
 	editable: boolean;
-	prePopulate?: () => void;
+	setEditable?: React.Dispatch<React.SetStateAction<boolean>>;
+	placeholderText?: string;
+	htmlData?: string;
+	hideToolbar?: boolean;
+	showToolbarText?: boolean;
 };
 
 const TextEditor = ({
+	mentionsData = false,
 	useMentionLookupService,
-	convertFileToImageUrl,
+	convertFilesToImageUrl,
 	onChangeCallback,
 	wrapperClass,
 	editable,
-	prePopulate,
+	setEditable,
+	placeholderText,
+	htmlData,
+	hideToolbar = false,
+	showToolbarText = false,
 }: Props) => {
 	const editorRef = useRef<LexicalEditor>(null);
+
 	const initialConfig = {
-		editorState: prePopulate || null,
 		namespace: "Editor",
 		theme: EditorTheme,
 		onError: (error: Error) => {
@@ -60,11 +81,17 @@ const TextEditor = ({
 				<TableContext>
 					<div className="editor-shell">
 						<Editor
-							convertFileToImageUrl={convertFileToImageUrl}
+							htmlData={htmlData}
+							convertFilesToImageUrl={convertFilesToImageUrl}
+							mentionsData={mentionsData}
 							useMentionLookupService={useMentionLookupService}
 							onChangeCallback={onChangeCallback}
 							editState={editable}
+							setEditable={setEditable}
 							editorRef={editorRef}
+							placeholderText={placeholderText}
+							hideToolbar={hideToolbar}
+							showToolbarText={showToolbarText}
 						/>
 					</div>
 				</TableContext>