+++ development/text-editor-app/components/text-editor/Example/Example.tsx
@@ -7,19 +7,19 @@ const Example = () => {
 	const [editable, setEditable] = useState<boolean>(false);
 	const [data, setData] = useState<string | TrustedHTML>("Here");
 
-	function useMentionLookupService() {
-		return [
-			{ mentionName: "user_xxx", label: "xxx" },
-			{ mentionName: "user_yyy", label: "yyy" },
-		];
-	}
-
-	function convertFileToImageUrl(files: FileList | null) {
-		if (files) {
-			const imgUrl = URL.createObjectURL(files[0]);
-			return imgUrl;
+	function convertFilesToImageUrl(files: FileList | null) {
+		if (!files || files.length === 0) {
+			return null;
 		}
-		return null;
+		const imageUrls = [];
+
+		for (let i = 0; i < files.length; i++) {
+			const file = files[i];
+
+			const imageUrl = URL.createObjectURL(file);
+			imageUrls.push(imageUrl);
+		}
+		return imageUrls.length > 0 ? imageUrls : null;
 	}
 
 	function onChange(_editorRef: LexicalEditor | null, payload: any) {
@@ -30,20 +30,13 @@ const Example = () => {
 		<div className="flex flex-col items-center justify-center py-10 mt-10 bg-gray-100 border rounded gap-x-3">
 			<TextEditor
 				editable={editable}
-				useMentionLookupService={useMentionLookupService}
-				convertFileToImageUrl={convertFileToImageUrl}
+				setEditable={setEditable}
+				convertFilesToImageUrl={convertFilesToImageUrl}
 				onChangeCallback={onChange}
+				hideToolbar={false}
+				showToolbarText={false}
 			/>
-			{editable ? (
-				<button
-					className="button"
-					onClick={() => {
-						setEditable(false);
-					}}
-				>
-					Save
-				</button>
-			) : (
+			{!editable && (
 				<div
 					className="w-full min-h-[50px] cursor-text bg-white"
 					dangerouslySetInnerHTML={{ __html: data }}