+++ development/text-editor-app/components/text-editor/PrepopulatedTextEditor/PrepopulatedTextEditor.jsx
@@ -1,59 +1,87 @@
-import React, { useState } from "react";
-import { $getRoot, $createTextNode, $createParagraphNode } from "lexical";
+import React, { useEffect, useState } from "react";
 import { TextEditor } from "@locoworks/reusejs-react-text-editor";
 import "@locoworks/reusejs-react-text-editor/css";
 
 const PrepopulatedTextEditor = () => {
+	const payload = `<table>
+      <tr>
+        <th>Company</th>
+        <th>Contact</th>
+        <th>Country</th>
+      </tr>
+      <tr>
+        <td>Alfreds Futterkiste</td>
+        <td>Maria Anders</td>
+        <td>Germany</td>
+      </tr>
+      <tr>
+        <td>Centro comercial Moctezuma</td>
+        <td>Francisco Chang</td>
+        <td>Mexico</td>
+      </tr>
+      <tr>
+        <td>Ernst Handel</td>
+        <td>Roland Mendel</td>
+        <td>Austria</td>
+      </tr>
+      <tr>
+        <td>Island Trading</td>
+        <td>Helen Bennett</td>
+        <td>UK</td>
+      </tr>
+      <tr>
+        <td>Laughing Bacchus Winecellars</td>
+        <td>Yoshi Tannamuri</td>
+        <td>Canada</td>
+      </tr>
+      <tr>
+        <td>Magazzini Alimentari Riuniti</td>
+        <td>Giovanni Rovelli</td>
+        <td>Italy</td>
+      </tr>
+   </table>`;
+	const [htmlData, setHtmlData] = useState(`<p>Here hello</p>`);
+
+	setTimeout(() => {
+		setHtmlData(payload);
+	}, 5000);
+
+	useEffect(() => {
+		setData(htmlData);
+	}, [htmlData]);
+
 	const [editable, setEditable] = useState(false);
-	const [data, setData] = useState("");
+	const [data, setData] = useState(htmlData);
 
-	function useMentionLookupService() {
-		return [
-			{ mentionName: "user_xxx", label: "xxx" },
-			{ mentionName: "user_yyy", label: "yyy" },
-		];
-	}
-
-	function convertFileToImageUrl(files) {
-		if (files) {
-			const imgUrl = URL.createObjectURL(files[0]);
-			return imgUrl;
+	function convertFilesToImageUrl(files) {
+		if (!files || files.length === 0) {
+			return null;
 		}
-		return null;
+		const imageUrls = [];
+
+		for (let i = 0; i < files.length; i++) {
+			const file = files[i];
+
+			const imageUrl = URL.createObjectURL(file);
+			imageUrls.push(imageUrl);
+		}
+		return imageUrls.length > 0 ? imageUrls : null;
 	}
 
 	function onChange(_editorRef, payload) {
 		setData(payload["html"]);
 	}
 
-	function prepopulatedRichText() {
-		const root = $getRoot();
-		if (root.getFirstChild() === null) {
-			const paragraph = $createParagraphNode();
-			paragraph.append($createTextNode("Start here!"));
-			root.append(paragraph);
-		}
-	}
-
 	return (
 		<div className="flex flex-col items-center justify-center py-10 mt-10 bg-gray-100 border rounded gap-x-3">
 			<TextEditor
-				prePopulate={prepopulatedRichText}
+				htmlData={htmlData}
 				editable={editable}
-				useMentionLookupService={useMentionLookupService}
-				convertFileToImageUrl={convertFileToImageUrl}
+				setEditable={setEditable}
+				convertFilesToImageUrl={convertFilesToImageUrl}
 				onChangeCallback={onChange}
 			/>
-			{editable ? (
-				<button
-					className="button"
-					onClick={() => {
-						setEditable(false);
-					}}
-				>
-					Save
-				</button>
-			) : (
+			{!editable && (
 				<div
 					className="w-full min-h-[50px] cursor-text bg-white"
 					dangerouslySetInnerHTML={{ __html: data }}