+++ components/date-picker/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @locoworks/reusejs-react-date-picker
 
+## 3.1.1
+
+### Patch Changes
+
+- 9ad33a7: Added support for empty default date
+
+## 3.1.0
+
+### Minor Changes
+
+- 3d6ce17: -Added Month Selector Component
+  -Added Year Selector Component
+  and related props
+
 ## 3.0.2
 
 ### Patch Changes